Move uuid generator to utils
